+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [3.56.0] - 2024-03-26
 
+### Added
+- Config option `disableStorageApi` to prevent `localStorage` operations
+
 ### Changed
 - `localStorage` availability check to not create a test-entry anymore